+++ ZeneFit/Data/Repository/Home/HomePepository.swift
@@ -1,0 +1,12 @@
+//
+//  HomePepository.swift
+//  ZeneFit
+//
+//  Created by iOS신상우 on 11/12/23.
+//
+
+import Foundation
+
+final class HomePepository {
+    
+}
