+++ ZeneFit/Domain/Model/Auth/SignUpInfo.swift
