+++ ZeneFit/Domain/Model/Policy/RecommendWelFareEntity.swift
