+++ ZeneFit/Domain/Repository/HomeRepositoryProtocol.swift
@@ -1,0 +1,12 @@
+//
+//  HomeRepositable.swift
+//  ZeneFit
+//
+//  Created by iOS신상우 on 11/12/23.
+//
+
+import Foundation
+
+protocol HomeRepositoryProtocol {
+    
+}
